Seperate Hosting Infrastructure and Api
