+++ Backend/src/SSAH.Infrastructure.Api/DependencyRegistry.cs
@@ -7,8 +7,5 @@ namespace SSAH.Infrastructure.Api
         public static void Configure(ContainerBuilder builder)
         {
         }
-
-        // TODO: Remove
-        public static IContainer Container { get; set; }
     }
 }
