+++ Backend/src/SSAH.Infrastructure.Api/SSAH.Infrastructure.Api.csproj
@@ -7,7 +7,8 @@
 
   <ItemGroup>
     <PackageReference Include="Autofac" Version="4.7.0" />
-    <PackageReference Include="Microsoft.AspNetCore.Mvc.Core" Version="2.0.3" />
+    <PackageReference Include="Microsoft.AspNetCore.Diagnostics" Version="2.0.2" />
+    <PackageReference Include="Microsoft.AspNetCore.Mvc" Version="2.0.3" />
     <PackageReference Include="Microsoft.Extensions.Logging.Abstractions" Version="2.0.1" />
     <PackageReference Include="Microsoft.Extensions.Options" Version="2.0.1" />
   </ItemGroup>
