+++ Backend/src/SSAH.Startup/Startup.cs
@@ -1,42 +1,23 @@
-﻿using Microsoft.AspNetCore.Builder;
+﻿using Autofac;
+
+using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Hosting;
-using Microsoft.AspNetCore.Mvc.Controllers;
-using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
-using Microsoft.Extensions.DependencyInjection.Extensions;
-using Microsoft.Extensions.Logging;
 
 using SSAH.Infrastructure.Api;
-using SSAH.Infrastructure.Api.Controllers;
-using SSAH.Infrastructure.Api.Pipeline;
 
 namespace SSAH.Startup
 {
     public class Startup
     {
-        public Startup(IConfiguration configuration)
-        {
-            Configuration = configuration;
-        }
-
-        public IConfiguration Configuration { get; }
-
-        // This method gets called by the runtime. Use this method to add services to the container.
         public void ConfigureServices(IServiceCollection services)
         {
-            services.AddMvc().Services.Replace(ServiceDescriptor.Transient<IControllerActivator, UnitOfWorkControllerActivator>());
+            services.AddSnowSchoolAdministrationHub();
         }
 
-        // This method gets called by the runtime. Use this method to configure the HTTP request pipeline.
-        public void Configure(IApplicationBuilder app, IHostingEnvironment env)
+        public void Configure(IApplicationBuilder app, IHostingEnvironment env, IContainer rootContainer)
         {
-            if (env.IsDevelopment())
-            {
-                // app.UseDeveloperExceptionPage();
-            }
-
-            app.UseSnowSchoolAdministrationHub(DependencyRegistry.Container);
-            app.UseMvc();
+            app.UseSnowSchoolAdministrationHub(env, rootContainer);
         }
     }
 }
